hubble-relay: Send node status events in GetFlows
This adds support for NodeStatusEvents in the GetFlows RPC call.
The intent of this event is to inform downstream consumers about the
state of the nodes which are participating in the current request.

The node state `NODE_GONE` is not used in this initial version, as the
current implementation of the peer management in hubble-relay does not
inform the running GetFlows request about the removal of a node.

If a peer is not ready when the request is started, we mark it as
`NODE_UNAVAILABLE`. If a peer errors out during a request, we indicate
this via `NODE_ERROR` and propagate the received error message.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>