k8s: Fix CCNP for host policies
This commit fixes the CNP/CCNP parsing to accept NodeSelectors, used to
define host policies. NodeSelectors are only accepted in CCNP resources.
Because host policies are never limited to a namespace, it would be
misleading to accept NodeSelectors in CNP resources.

This commit also adds a new CCNP example of host policies in the same
directory as the existing JSON example.

Fixes: f9c205d ("pkg/policy: Host network policies")
Signed-off-by: Paul Chaignon <paul@cilium.io>